% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{flanker}
\alias{flanker}
\title{Data: Flanker Task data from \insertCite{hedge2018reliability;textual}{vICC}.}
\format{
A dataframe 33660 rows and 7 columns.
}
\usage{
data("flanker")
}
\description{
A dataset containing 33660 rows and 7 columns.

\itemize{
\item Block
\item Trial number
\item Arrow direction (1=left, 2=right)
\item Condition (0 = congruent, 1=neutral, 2=incongruent)
\item Correct (1) or incorrect (0)
\item Reaction time (seconds)
}
}
\note{
Reaction times less than 0.20 and greater than 2 seconds were removed.
}
\references{
\insertAllCited{}
}
\keyword{datasets}
