% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pip.R
\name{pip}
\alias{pip}
\title{Posterior Inclusion Probabilities}
\usage{
pip(object, ...)
}
\arguments{
\item{object}{Ab object of class \code{vicc}.}

\item{...}{Currently ignored.}
}
\value{
A data frame
}
\description{
Extract the posterior inclusion probabilities for either the
             random intercepts for sigma or the random effects standard deviation
             for sigma.
}
\examples{
\donttest{

# congruent trials
congruent <- subset(flanker, cond == 0)

# subset 25 from each group
dat <- congruent[unlist(tapply(1:nrow(congruent),
                            congruent$id,
                            head, 25)), ]

# fit model
fit <- vicc(y  = dat$rt,
            group = dat$id,
            iter = 250,
            burnin = 10,
            type =  "pick_group")

pip(fit)
}
}
