% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pip.R
\name{print.pip}
\alias{print.pip}
\title{Print \code{pip} Objects}
\usage{
\method{print}{pip}(x, ...)
}
\arguments{
\item{x}{An object of class \code{pip}.}

\item{...}{Currently ignored.}
}
\description{
Print \code{pip} Objects
}
