% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vicc.R
\name{print.vicc}
\alias{print.vicc}
\title{Print \code{vicc} Objects}
\usage{
\method{print}{vicc}(x, cred = 0.95, ...)
}
\arguments{
\item{x}{An object of class \code{vicc}.}

\item{cred}{Numeric. Credible interval width (defaults to \code{0.90}).}

\item{...}{Currently ignored}
}
\description{
Print \code{vicc} Objects
}
