% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sem.R
\name{sem}
\alias{sem}
\title{Standard Error of Measurement}
\usage{
sem(object, cred = 0.95, ...)
}
\arguments{
\item{object}{An object of class \code{vicc}.}

\item{cred}{Numeric. Credible interval width (defaults to \code{0.90}).}

\item{...}{Currently ignored.}
}
\value{
An object of class \code{sem}.
}
\description{
Compute the standard error of measurment, possibly for
             each group.
}
\note{
Due to the hierarchical model formulation, the group-level sem is
      partially pooled (i.e., shrinkage) to the average,
}
\examples{
\donttest{
# congruent trials
congruent <- subset(flanker, cond == 0)

# subset 25 from each group
dat <- congruent[unlist(tapply(1:nrow(congruent),
                            congruent$id,
                            head, 25)), ]

fit <- vicc(
  y  = dat$rt,
  group = dat$id,
  iter = 500,
  burnin = 10,
  type =  "pick_group"
  )

sem(fit)
}
}
