% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vicc.R
\name{vicc}
\alias{vicc}
\title{Varying Intraclass Correlation Coefficients}
\usage{
vicc(
  y,
  group,
  model = "pick_group",
  iter = 5000,
  chains = 2,
  burnin = 500,
  prior_scale = 1,
  prior_prob = 0.5
)
}
\arguments{
\item{prior_prob}{}
}
\value{

}
\description{
Varying Intraclass Correlation Coefficients
}
